refactor: split cli-add and flatten prompt retries
