+++ packages/wp-typia-project-tools/src/runtime/cli-add-shared.ts
@@ -1,0 +1,344 @@
+import fs from "node:fs";
+import { promises as fsp } from "node:fs";
+import path from "node:path";
+
+import {
+	HOOKED_BLOCK_ANCHOR_PATTERN,
+	HOOKED_BLOCK_POSITION_IDS,
+	type HookedBlockPositionId,
+} from "./hooked-blocks.js";
+import {
+	toKebabCase,
+	toSnakeCase,
+} from "./string-case.js";
+import {
+	type WorkspaceInventory,
+} from "./workspace-inventory.js";
+import {
+	WORKSPACE_TEMPLATE_PACKAGE,
+	type WorkspaceProject,
+} from "./workspace-project.js";
+
+/**
+ * Supported top-level `wp-typia add` kinds exposed by the canonical CLI.
+ */
+export const ADD_KIND_IDS = ["block", "variation", "pattern", "binding-source", "hooked-block"] as const;
+export type AddKindId = (typeof ADD_KIND_IDS)[number];
+
+/**
+ * Supported built-in block families accepted by `wp-typia add block --template`.
+ */
+export const ADD_BLOCK_TEMPLATE_IDS = [
+	"basic",
+	"interactivity",
+	"persistence",
+	"compound",
+] as const;
+export type AddBlockTemplateId = (typeof ADD_BLOCK_TEMPLATE_IDS)[number];
+
+const WORKSPACE_GENERATED_SLUG_PATTERN = /^[a-z][a-z0-9-]*$/;
+
+export interface RunAddVariationCommandOptions {
+	blockName: string;
+	cwd?: string;
+	variationName: string;
+}
+
+export interface RunAddPatternCommandOptions {
+	cwd?: string;
+	patternName: string;
+}
+
+export interface RunAddBindingSourceCommandOptions {
+	bindingSourceName: string;
+	cwd?: string;
+}
+
+export interface RunAddHookedBlockCommandOptions {
+	anchorBlockName: string;
+	blockName: string;
+	cwd?: string;
+	position: string;
+}
+
+export interface RunAddBlockCommandOptions {
+	blockName: string;
+	cwd?: string;
+	dataStorageMode?: string;
+	persistencePolicy?: string;
+	templateId?: string;
+}
+
+export interface WorkspaceMutationSnapshot {
+	fileSources: Array<{
+		filePath: string;
+		source: string | null;
+	}>;
+	snapshotDirs: string[];
+	targetPaths: string[];
+}
+
+export function normalizeBlockSlug(input: string): string {
+	return toKebabCase(input);
+}
+
+export function assertValidGeneratedSlug(label: string, slug: string, usage: string): string {
+	if (!slug) {
+		throw new Error(`${label} is required. Use \`${usage}\`.`);
+	}
+	if (!WORKSPACE_GENERATED_SLUG_PATTERN.test(slug)) {
+		throw new Error(
+			`${label} must start with a letter and contain only lowercase letters, numbers, and hyphens.`,
+		);
+	}
+
+	return slug;
+}
+
+export function assertValidHookedBlockPosition(position: string): HookedBlockPositionId {
+	if ((HOOKED_BLOCK_POSITION_IDS as readonly string[]).includes(position)) {
+		return position as HookedBlockPositionId;
+	}
+
+	throw new Error(
+		`Hook position must be one of: ${HOOKED_BLOCK_POSITION_IDS.join(", ")}.`,
+	);
+}
+
+export function getWorkspaceBootstrapPath(workspace: WorkspaceProject): string {
+	const workspaceBaseName = workspace.packageName.split("/").pop() ?? workspace.packageName;
+	return path.join(workspace.projectDir, `${workspaceBaseName}.php`);
+}
+
+export function buildWorkspacePhpPrefix(workspacePhpPrefix: string, slug: string): string {
+	return toSnakeCase(`${workspacePhpPrefix}_${slug}`);
+}
+
+export function isAddBlockTemplateId(value: string): value is AddBlockTemplateId {
+	return (ADD_BLOCK_TEMPLATE_IDS as readonly string[]).includes(value);
+}
+
+export function quoteTsString(value: string): string {
+	return JSON.stringify(value);
+}
+
+export async function patchFile(
+	filePath: string,
+	transform: (source: string) => string,
+): Promise<void> {
+	const currentSource = await fsp.readFile(filePath, "utf8");
+	const nextSource = transform(currentSource);
+	if (nextSource !== currentSource) {
+		await fsp.writeFile(filePath, nextSource, "utf8");
+	}
+}
+
+export async function readOptionalFile(filePath: string): Promise<string | null> {
+	if (!fs.existsSync(filePath)) {
+		return null;
+	}
+
+	return fsp.readFile(filePath, "utf8");
+}
+
+export async function restoreOptionalFile(filePath: string, source: string | null): Promise<void> {
+	if (source === null) {
+		await fsp.rm(filePath, { force: true });
+		return;
+	}
+
+	await fsp.mkdir(path.dirname(filePath), { recursive: true });
+	await fsp.writeFile(filePath, source, "utf8");
+}
+
+export async function snapshotWorkspaceFiles(
+	filePaths: string[],
+): Promise<WorkspaceMutationSnapshot["fileSources"]> {
+	const uniquePaths = Array.from(new Set(filePaths));
+	return Promise.all(
+		uniquePaths.map(async (filePath) => ({
+			filePath,
+			source: await readOptionalFile(filePath),
+		})),
+	);
+}
+
+export async function rollbackWorkspaceMutation(snapshot: WorkspaceMutationSnapshot): Promise<void> {
+	for (const targetPath of snapshot.targetPaths) {
+		await fsp.rm(targetPath, { force: true, recursive: true });
+	}
+	for (const snapshotDir of snapshot.snapshotDirs) {
+		await fsp.rm(snapshotDir, { force: true, recursive: true });
+	}
+	for (const { filePath, source } of snapshot.fileSources) {
+		await restoreOptionalFile(filePath, source);
+	}
+}
+
+export function resolveWorkspaceBlock(
+	inventory: WorkspaceInventory,
+	blockSlug: string,
+): WorkspaceInventory["blocks"][number] {
+	const block = inventory.blocks.find((entry) => entry.slug === blockSlug);
+	if (!block) {
+		throw new Error(
+			`Unknown workspace block "${blockSlug}". Choose one of: ${inventory.blocks.map((entry) => entry.slug).join(", ")}`,
+		);
+	}
+	return block;
+}
+
+export function assertValidHookAnchor(anchorBlockName: string): string {
+	const trimmed = anchorBlockName.trim();
+	if (!trimmed) {
+		throw new Error(
+			"`wp-typia add hooked-block` requires --anchor <anchor-block-name>.",
+		);
+	}
+	if (!HOOKED_BLOCK_ANCHOR_PATTERN.test(trimmed)) {
+		throw new Error(
+			"`wp-typia add hooked-block` requires --anchor <anchor-block-name> to use the full `namespace/slug` block name format.",
+		);
+	}
+
+	return trimmed;
+}
+
+export function readWorkspaceBlockJson(
+	projectDir: string,
+	blockSlug: string,
+): {
+	blockJson: Record<string, unknown>;
+	blockJsonPath: string;
+} {
+	const blockJsonPath = path.join(projectDir, "src", "blocks", blockSlug, "block.json");
+	if (!fs.existsSync(blockJsonPath)) {
+		throw new Error(
+			`Missing ${path.relative(projectDir, blockJsonPath)} for workspace block "${blockSlug}".`,
+		);
+	}
+
+	let blockJson: unknown;
+	try {
+		blockJson = JSON.parse(fs.readFileSync(blockJsonPath, "utf8"));
+	} catch (error) {
+		throw new Error(
+			error instanceof Error
+				? `Failed to parse ${path.relative(projectDir, blockJsonPath)}: ${error.message}`
+				: `Failed to parse ${path.relative(projectDir, blockJsonPath)}.`,
+		);
+	}
+
+	if (!blockJson || typeof blockJson !== "object" || Array.isArray(blockJson)) {
+		throw new Error(
+			`${path.relative(projectDir, blockJsonPath)} must contain a JSON object.`,
+		);
+	}
+
+	return {
+		blockJson: blockJson as Record<string, unknown>,
+		blockJsonPath,
+	};
+}
+
+export function getMutableBlockHooks(
+	blockJson: Record<string, unknown>,
+	blockJsonRelativePath: string,
+): Record<string, string> {
+	const blockHooks = blockJson.blockHooks;
+	if (blockHooks === undefined) {
+		const nextHooks: Record<string, string> = {};
+		blockJson.blockHooks = nextHooks;
+		return nextHooks;
+	}
+	if (!blockHooks || typeof blockHooks !== "object" || Array.isArray(blockHooks)) {
+		throw new Error(`${blockJsonRelativePath} must define blockHooks as an object when present.`);
+	}
+
+	return blockHooks as Record<string, string>;
+}
+
+export function assertVariationDoesNotExist(
+	projectDir: string,
+	blockSlug: string,
+	variationSlug: string,
+	inventory: WorkspaceInventory,
+): void {
+	const variationPath = path.join(
+		projectDir,
+		"src",
+		"blocks",
+		blockSlug,
+		"variations",
+		`${variationSlug}.ts`,
+	);
+	if (fs.existsSync(variationPath)) {
+		throw new Error(
+			`A variation already exists at ${path.relative(projectDir, variationPath)}. Choose a different name.`,
+		);
+	}
+	if (
+		inventory.variations.some(
+			(entry) => entry.block === blockSlug && entry.slug === variationSlug,
+		)
+	) {
+		throw new Error(
+			`A variation inventory entry already exists for ${blockSlug}/${variationSlug}. Choose a different name.`,
+		);
+	}
+}
+
+export function assertPatternDoesNotExist(
+	projectDir: string,
+	patternSlug: string,
+	inventory: WorkspaceInventory,
+): void {
+	const patternPath = path.join(projectDir, "src", "patterns", `${patternSlug}.php`);
+	if (fs.existsSync(patternPath)) {
+		throw new Error(
+			`A pattern already exists at ${path.relative(projectDir, patternPath)}. Choose a different name.`,
+		);
+	}
+	if (inventory.patterns.some((entry) => entry.slug === patternSlug)) {
+		throw new Error(
+			`A pattern inventory entry already exists for ${patternSlug}. Choose a different name.`,
+		);
+	}
+}
+
+export function assertBindingSourceDoesNotExist(
+	projectDir: string,
+	bindingSourceSlug: string,
+	inventory: WorkspaceInventory,
+): void {
+	const bindingSourceDir = path.join(projectDir, "src", "bindings", bindingSourceSlug);
+	if (fs.existsSync(bindingSourceDir)) {
+		throw new Error(
+			`A binding source already exists at ${path.relative(projectDir, bindingSourceDir)}. Choose a different name.`,
+		);
+	}
+	if (inventory.bindingSources.some((entry) => entry.slug === bindingSourceSlug)) {
+		throw new Error(
+			`A binding source inventory entry already exists for ${bindingSourceSlug}. Choose a different name.`,
+		);
+	}
+}
+
+/**
+ * Returns help text for the canonical `wp-typia add` subcommands.
+ */
+export function formatAddHelpText(): string {
+	return `Usage:
+  wp-typia add block <name> --template <${ADD_BLOCK_TEMPLATE_IDS.join("|")}> [--data-storage <post-meta|custom-table>] [--persistence-policy <authenticated|public>]
+  wp-typia add variation <name> --block <block-slug>
+  wp-typia add pattern <name>
+  wp-typia add binding-source <name>
+  wp-typia add hooked-block <block-slug> --anchor <anchor-block-name> --position <${HOOKED_BLOCK_POSITION_IDS.join("|")}>
+
+Notes:
+  \`wp-typia add\` runs only inside official ${WORKSPACE_TEMPLATE_PACKAGE} workspaces.
+  \`add variation\` targets an existing block slug from \`scripts/block-config.ts\`.
+  \`add pattern\` scaffolds a namespaced PHP pattern shell under \`src/patterns/\`.
+  \`add binding-source\` scaffolds shared PHP and editor registration under \`src/bindings/\`.
+  \`add hooked-block\` patches an existing workspace block's \`block.json\` \`blockHooks\` metadata.`;
+}
